Fix PDF build error with Texinfo in the unusual location
When texinfo is not in the /usr/bin, but in the custom location, then texi2pdf
will fail to grab texinfo.tex macros file from its own distribution. At the
same time, newlib/lib{c,m}/Makefile.in has an error where path to the
texinfo.tex, shipped with newlib is invalid - it does not have enough ".."
entries in it. To avoid this problem we have to explicitly set TEXINPUTS
variable to point to texinfo.tex from newlib repository.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>