Fix sporadic build error when installing libssp headers
It looks like that libssp install target is not parallel-friendly - I had
occassional issues, when installing it's header. I'm not entirely sure that
this is the case, but that is my best guess so far, considering that I
wasn't able to see any particular pattern of when it does happen. Error
message in log file is:

/usr/bin/install: cannot create regular file
`/home/arc/arc_gnu/toolchain/../INSTALL/lib/gcc/arc-elf32/6.1.1/include/ssp/stdio.h':
File exists

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>