Build GDB as a C application
Due to STAR 9001066513 GDB crashes when throwing an exception.  Due to a poor
design GDB uses exceptions as a part of a normal execution flow to verify user
input (for example it throws exception when file path is invalid or when typed
symbol doesn't exist).  Therefore for 2016.09 release it is required to disable
build of GDB as a C++ application.  This will not work for future releases,
because GDB will stop supporting C builds.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>