uClibc: Build uClibc outside of source tree
Historically it wasn't possible to build uClibc out of the source tree. As a
results our build scripts had to build right in the source tree and to
copy sources when using custom build tree location. I'm not sure was were
the errors in the past, but at least as of uClibc-ng 1.0.17 the only missing
fix is a small change to extra/locale/Makefile.in file. With that fix in
place it is now possible to build uClibc out of the source tree, and
therefore there is no need to copy sources to build tree in some occasions.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>