Build GDB for windows as a C application
Similar to ARC Linux, on Windows with Mingw there is a bug, where GDB, built
as a C++ application, sometimes crashes when an exception is thrown in it.
It seems that behaviour is somewhat different from ARC Linux case, because
on Windows it seems that it doesn't crash in some cases where it crashes on
ARC Linux.

As a workaround, for now built GDB as a C application to avoid using
exceptions.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>