uClibc: Do not configure uClibc twice
For some reason currently build-uclibc.sh configures uClibc twice: first
when installing headers, then when actually building uClibc. The only
difference in those configurations is `HARDWIRED_ABSPATH=n` which is set in
the second time. It seems that it is safe to configure uClibc just once
before header installation, then reuse same configuration when building it.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>