uClibc: Do not create temp defconfig file
To create uClibc defconfig without modifying source file script build-uclibc.sh
used to create temporary file that is a copy of defconfig, then modify it, then
use standard "defconfig" command, then remove temporary file. This looks like
an unnecessary complication considering that "olddefconfig" yields same result.
So this patch first creates a .config file which is really just a copy of a
specified defconfig, modifies it the same way as before, the uses `make
olddefconfig` to convert it to a full .config file.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>