Doc: Refresh script to build documentation
With this change doc/generate now will not assume that it can use PIP -
instead it is expected that Sphinx and dependencies are alredy installed in
some custom location specified via environment variable.

This patch also removes assumption that texlive-texmf-latex package is
installed and instead relies on already preexisting directory that would
contain it.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>