Properly check IS_CROSS_COMPILING variable
Two checks for IS_CROSS_COMPILING value in build-elf32.sh assumed that this
variable is "true" if set, which is wrong, because build-all.sh always sets
this variable, either to "yes" or "no". This patch fixes this problem, now
checks is done against "yes" value of this variable to figure out of this is
a cross-compilation of toolchain.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>