Release: Strip -release from the IDE installer name
There is some complication in tag names and release names, because for
engineering builds and release candidates release names are same as tag
names, but without `arc-` prefix, but for final GA release tags there is
also a suffix `-release` that should be removed from the release name. This
suffix wasn't stripped in windows installer script, which is corrected in
this patch.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>