Add --uclibc-build-in-src-tree option
Add an option that forces build script to build uClibc inside the source
tree - the same way how it was done before. Following factors collide
together:

* In out of tree build path to uClibc build directory has to be absolute;
* During final link stages all files are explicitly specified to linker or
  archiver.
* In automated build/test environment it is common to have very long paths.

Combined together those factors create an argument list that is bigger than
allowed by Linux, so build fails.  When build is done in the source tree
then paths are relative so issue doesn't happen.

This patch allows user to control if build is done in source tree or not to
adapt to environment.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>