Release: Ensure that UTF-8 encoding is used
There was a bug, where git log generation was failing if non-ASCII symbols
were present in the stream. This was happening, because system is setup to
use C locale be default instead of en_US.utf8, so Python was trying to use
ASCII codec to decode git output with non-ASCII symbols. To avoid problems
and dependencies on external environment, this patch makes sure that Python
script always uses UTF-8.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>