Release: Explicitly disable libusb-0.1 usage for OpenOCD
OpenOCD configure script uses pkg-config to identify if libusb-0.1 and
libusb-1.0 are present.  This causes issues with cross builds on MinGW,
because pkg-config gets confused and assumes that host libusb-0.1 is ok for
MinGW build, which is not true.  One possible solution to this problem is to
disable usage of pkg-config and instead directly pass -L and -I options
required to use libusb-1.0.  However this causes another trouble, as OpenOCD
has some "hack" for libusb bug, and when libusb-1.0 is supplied directly it
assumes that version <1.0.9 is used, but that causes a compatibility issues
with libusb-1.0.20 that Synopsys uses in GNU Toolchain builds.

As a solution and option --disable-libusb0 has been added to OpenOCD configure
scripts that allows to disable libusb0 usage at all, since it is not needed for
Synopsys builds.  Default behavior is unchanged (use pkg-config to search for
libusb-0.1), but if option is specified, then its value is used.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>