Release: Set PKG_CONFIG_PATH as argument to ./configure
In OpenOCD both configure and configure.ac come from Git, hence may have
same timestamp and as a result, makefile might try to regenerate configure.
This might fail when cross building for Windows, because PKG_CONFIG_PATH
shall be set properly to specify path to libusb-1.0 built for Windows hosts.
But in this second run it is not set, hence reconfiguration fails. The
solution is that this variable should be set not as an envvar to
./configure, but as an argument to it, thus it would be logged and set
properly on the reconfigure runs. Hence, this:

	PKG_CONFIG_PATH=... ./configure

is replaced with this:

	./configure PKG_CONFIG_PATH=...

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>