Release: Use SHA-256 instead of MD5 for check sums
MD5 is an insecure hash, which can be used to protect only from
unintentional file corruption. While that in general is ok for our GNU
releases, since we don't expect people trying to replace them with
anything malicious, I feel that it is time to move to a more secure hash,
considering that Synopsys makes a lot of effort to increase security of its
products.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>