Fix issues with uClibc sources checkout in engineering builds
The commit, that added support for glibc changed behaviour of
arc-versions.sh script. Previously this script would checkout uClibc sources
at all occasions, but after this commit it would checkout uClibc sources
only when uclibc toolchain is being built; it would checkout glibc for glibc
toolchain; it wouldn't checkout neither uClibc, not glibc for baremetal
toolchain. This is an intended behaviour, since there is really no reason to
enforce uClibc onto those who build baremetal toolchain.

This had an adverse effect on the process of toolchain releases/engineering
builds though, where release.mk calls tag-release.sh, which in
turn invokes arc-versions.sh. There are actually two simultaneous problems
here.

First is that tag-release.sh doesn't set neither DO_UCLIBC or DO_GLIBC, so
the new code fails due to syntax error. This can be easily fixed in
arc-versions.sh, by setting default values for those variables in use cases.

Second is that when doing release in the future, with glibc, arc-versions.sh
would need to checkout both uClibc and glibc, which it doesn't support right
now.

Third problem is that both tag-release.sh and build-all.sh might call
arc-version.sh to checkout the sources without doing the build, and in case
of build-all.sh that would mean that both DO_UCLIBC and DO_GLIBC are set to
"no", so build-all.sh would need to set them to "yes" for the case when all
of the DO_ELF32, DO_UCLIBC and DO_GLIBC are set to "no" - but that should be
done only for arc-versions.sh invocation.

This patch fixes only the first problem, which is enough for now, when
releasables are the same as they used to be. Second and third problems
should be addressed in the future when it would be needed to add glibc to
the release.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>