readme: Specify that whitespaces are not supported
Back in the day build scripts where written in a way to support paths with
whitespaces - all filenames and directories were surrounded by the quotes,
but there are reports that this doesn't work anymore and whitespaces break
build process, which is hardly a surprise because I haven't tried to
maintain this feature. However, limitation of "no whitespace" has never been
documented in README and this commit corrects this. Perhaps even better
would be do add some check at the top of build-all.sh?

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>