Don't use uClibc's defconfig for ARC
uClibc at some point (I don't remember exact version) removed built-in
defconfigs for architectures, instead whoever builds the toolchain should
provide their own defconfig.  This patch adds compatibility with those
versions of uClibc, by including a copy of defconfig that was in the uClibc
for ARC, and by using it. Code changes are pretty minimal.  uClibc
configurations still can be selected via --uclibc-defconfig option, as it
was before, but now defconfig to select should be in the
toolchain/extras/uClibc directory.

Another change addressed in this patch is that uClibc now doesn't select
between ARC 700 and ARC HS via it's own options and instead uses whatever is
the default CPU of the compiler, thus there is not difference between ARC HS
and ARC 700 configurations and single config can be used.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>