Add patch that enables gcc plugins on win32
By default it is not possible to build GCC with plugins support for Windows
hosts when using Mingw cross-toolchain. Added patch to gcc addresses this
problem, but as of know it can't be applied to source tree unconditionally,
since it would break normal build on Linux. This patch is added to
repository so that it will not be lost in times, so those who need it, can
use it, but it would be good to find some generic solution for upstream
repository.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>