Refresh arc-linux-aa4.exp
Remove redundant comments from arc-linux-aa4.exp. This file also better be
renamed since it isn't really related to ArcAngel4 (most people don't even
know what this is), but this board is used by the name in verification, so
file can be renamed only together with synchronous update to verification.

It is actually quite possible that the best way is to merge this file into
arc-linux-generic.exp, because there is not much specifics in it.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>