Remove arc-linux-nsim.exp
This doesn't really make sense, because it is pretty much the same as
arc-linux-aa4. The only substantial difference is that arc-linux-aa4 uses a
fixed IP address of target specified via environment variable, while
arc-linux-nsim.exp uses a list of possible IP addresses from a file, so in
theory can use a pool of target machines. However IP address is selected
once at the start of DejaGNU testsuite, so it is possible to just select
machine from a pool at higher level and pass via variable to arc-linux-aa4.

This board is not used by our verification and is reduntant.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>