Sync uClibc config with buildroot
This commit synchronizes uClibc configuration that we use to build toolchain
with the one from buildroot, which enables more features.

Buildroot config is from [1], commit 5beb2bd908. This is not a copy-paste,
I've omitted some things in their config that seems dubious too me, like
usage of PIC-only object files.

One important difference is that UCLIBC_HAS_LIBINTL=y is not set here, but
is set in the Buildroot. The reason is that when libintl is enabled such
toolchain then fails to build selfhosting gcc. I don't know the reason for
this, so for now I just don't add this option to our configuration.

[1] https://git.buildroot.net/buildroot/tree/package/uclibc/uClibc-ng.config

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>