Fix absence of nano libs in windows builds
It has been observed that Windows releases of GNU Toolchain lack libraries
that are built with optimization for size. This happened because of the
invalid algorithm that was used to determine list of multilibs on target -
it was trying to use wrong absolute path to the target compiler. However in
case of cross compiler build this compiler is actually should be in the PATH
(by requirements of our scripts), so instead of futile attempts to calculate
abs path scripts should just use whatever is in the PATH.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>