Rename EM Starter Kit 2.03a to 2.3
That version of EM Starter Kit was named "2.03a" for some time, thus GNU
documentation named it as such. But later it was decided to name it 2.3
instead, but GNU docs were not updated accordingly at that time. This commit
fixes this and changes all mentions of 2.03a to 2.3.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>

e