doc: Remove dead link
There was a dead link at ide/installation.rst to
http://download.eclipse.org/tm/updates/3.7. In fact official TM site also
contains the very same link and it doesn't work as well, so it seems there
is no new version of this link. In fact, since this is a section related to
older GNU version (<=2016.03) it makes sense to just remove this whole
section completely.