linux: Build C++ compiler at stage 1
Glibc has dependency on C++ compiler, therefore ARC C++ compiler has to be built
already at stage 1. Building libstdc++ doesn't seem to be required, only
compiler is needed. This change is applied to both uClibc and Glibc
toolchains because they share common 'configure' invocation code, but this
is not really needed for uClibc.