release: Don't distribute JRE with macOS IDE
1. Eclipse wasn't really auto-recognizing JRE in the eclipse/jre folder, it
was requried to specify it explicitly in eclipse.ini file. This is,
presumably, result of assumption that Eclipse on macOS is distrubuted as a
.dmg file rather then .tar.gz.

2. We now use OpenJ9 as a VM, but I've found that as of version 11 it
doesn't work with Eclipse - there is an error at start. See
https://github.com/eclipse/openj9/issues/3795

Thus as a result for now we would assume that user has necessary Java
runtime is installed by the user.