doc: Avoid bogus and unused reference to morefloats package
According to [1] "morefloats" is needed to allow more floats (images) in the
PDF, however reading the code it becomes clear that in fact only one of
"float" and "morefloats" is used and experiment shows that is it "float"
that actually allows more images, rather then "morefloat". This commit
removes reference to "morefloats" to avoid confusion.

[1] 596bf47 doc: Increase limitations on the number of floats