Fix an issue with uclibc-ng tag in the release branch
uClibc-ng is an upstream repository, therefore tag to use should specify the
same tag, as in the arc-dev branch, rather then switch to arc release tag.
This wasn't changed when scripts switched to upstream repository, but until
now nobody has been noticing that. This commit fixes this issue, which was
found by RCT.