arc-versions: fix checking out repos when building without uclibc and glibc
This commit fixes an issue where arc-versions.sh would skip updating all
git repositories when building without uclibc and glibc. The check
responsible for this needs to be specific enough to only skip updating
the linux tree in that case, as indicated by the comment and similar
checks a few lines above.