dejagnu: increase timeout for tests execution on nsimdrv
Some tests are being executed for around 150 seconds (for example
crossmodule-in-dir-call-topn-1a.c). Thus sometimes they pass, sometimes
they fail due to timeout.

Also remove comment in code, because I think that it is quite obvious
that we setup timeout so that it should be more than longest test
execution time.