native_gdb: Specify the path to libgmp
On hosts that have their own libgmp installed, the check for libgmp
becomes something like:

$ arc-snps-linux-gnu-gcc -o conftest -lncurses -lm /usr/lib/libgmp.a

which will uses the hosts libgmp and not the native one. If there's
no libgmp installed on the host, then it would become:

$ arc-snps-linux-gnu-gcc -o conftest -lncurses -lm -l:libgmp.a

This commit adds the "--with-libgmp-prefix=" for configuring the
native_gdb so that we always pick the right library.