README.md: Added clarification on real build kick-start
For a seasoned CT-NG user it's all clear what to do and when, but for those who are just ramping-up on the new tool we need to be more explicit and precise on steps which need to be done. And indeed explicit command for the real build kick-start was missing. Add it now.