Revert "dejagnu: launch nsimdrv under time utility"
This reverts commit 6f7553762adf7cfd600bb778f1310469979cd8c0.

On a Debian machine without the "time" executable (/usr/bin/time),
the shell builtin is used which doesn't accept "-p" flag, which
leads to a misleading error in the log:

  Simulator time -p /path/to/nsim/bin/nsimdrv ... missing.
  FAIL: gcc.dg/20000623-1.c execution test

Since this change was only for debugging purposes, I don't see
the point to craft it even fruther like using "command -v time"
or optional "-p" flags, etc. Let's keep it simple.