arcv: Add support for riscv targets in nSIM setup.
Modified the nSIM setup logic to call setup_sim riscv when targeting
ARCV. It returns a simulator named with the target alias and a "-run"
suffix (for example, riscv64-unknown-elf-run), which can serve as a
wrapper for the simulator.

For non-ARCv targets, retain the previous behavior of setting the sim
board info directly.

Signed-off-by: Luis Silva <luiss@synopsys.com>